move graph generator to own file
