+++ data-generation/src/main/resources/reptilia-tortoise-network-sl.csv
@@ -1,0 +1,18 @@
+1 2 2012
+3 4 2012
+5 2 2012
+5 6 2012
+2 6 2012
+1 2 2013
+1 6 2013
+7 8 2013
+7 9 2013
+3 8 2013
+3 4 2013
+4 8 2013
+10 11 2013
+11 9 2013
+5 2 2013
+5 6 2013
+2 6 2013
+8 9 2013
